+++ ql/pricingengines/bond/all.hpp
@@ -3,4 +3,5 @@
 
 #include <ql/pricingengines/bond/bondfunctions.hpp>
 #include <ql/pricingengines/bond/discountingbondengine.hpp>
+#include <ql/pricingengines/bond/riskybondengine.hpp>
